_legacy/backend/requirements.txt
asgiref==3.10.0
dj-database-url==3.0.1
Django>=5.0,<6.0
djangorestframework==3.16.1
psycopg2-binary==2.9.11
python-dotenv==1.2.1
sqlparse==0.5.3
whitenoise==6.11.0
google-generativeai>=0.8.0
Pillow>=10.0.0
